feat: added WebViewBackButton and replaced references
